+++ worker/include/RTC/SCTP/association/NegotiatedCapabilities.hpp
@@ -15,33 +15,33 @@ namespace RTC
 		struct NegotiatedCapabilities
 		{
 			/**
+			 * Negotiated maximum number of outbound streams (OS).
+			 */
+			uint16_t maxOutboundStreams{ 0 };
+			/**
+			 * Negotiated maximum number of inbound streams (MIS).
+			 */
+			uint16_t maxInboundStreams{ 0 };
+			/**
 			 * Partial Reliability Extension.
 			 * @see RFC 3758.
 			 */
-			bool partialReliability = false;
+			bool partialReliability{ false };
 			/**
 			 * Stream Schedulers and User Message Interleaving (I-DATA Chunks).
 			 * @see RFC 8260.
 			 */
-			bool messageInterleaving = false;
+			bool messageInterleaving{ false };
 			/**
 			 * Stream Reconfiguration.
 			 * @see RFC 6525.
 			 */
-			bool reconfig = false;
+			bool reconfig{ false };
 			/**
 			 * Zero Checksum.
 			 * @see RFC 9653.
 			 */
-			bool zeroChecksum = false;
-			/**
-			 * Negotiated maximum number of outbound streams (OS).
-			 */
-			uint16_t maxOutboundStreams = 0;
-			/**
-			 * Negotiated maximum number of inbound streams (MIS).
-			 */
-			uint16_t maxInboundStreams = 0;
+			bool zeroChecksum{ false };
 
 			void Dump(int indentation = 0) const;
 		};