+++ CHANGELOG.md
@@ -2,8 +2,7 @@
 
 ### NEXT
 
-- Worker: Fix, encode retransmitted packets with the corresponding data
-  ([PR #1527](https://github.com/versatica/mediasoup/pull/1527)).
+- `Worker`: Fix encode retransmitted packets with the corresponding data ([PR #1527](https://github.com/versatica/mediasoup/pull/1527)).
 
 ### 3.15.7